requirements.txt
fastapi
jinja2
fpdf
python-dotenv
openai
uvicorn
